Reorganize util methods
