+++ homeassistant/util/color.py
@@ -1,0 +1,41 @@
+"""Color util methods."""
+
+
+# Taken from: http://www.cse.unr.edu/~quiroz/inc/colortransforms.py
+# License: Code is given as is. Use at your own risk and discretion.
+# pylint: disable=invalid-name
+def color_RGB_to_xy(R, G, B):
+    """ Convert from RGB color to XY color. """
+    if R + G + B == 0:
+        return 0, 0
+
+    var_R = (R / 255.)
+    var_G = (G / 255.)
+    var_B = (B / 255.)
+
+    if var_R > 0.04045:
+        var_R = ((var_R + 0.055) / 1.055) ** 2.4
+    else:
+        var_R /= 12.92
+
+    if var_G > 0.04045:
+        var_G = ((var_G + 0.055) / 1.055) ** 2.4
+    else:
+        var_G /= 12.92
+
+    if var_B > 0.04045:
+        var_B = ((var_B + 0.055) / 1.055) ** 2.4
+    else:
+        var_B /= 12.92
+
+    var_R *= 100
+    var_G *= 100
+    var_B *= 100
+
+    # Observer. = 2 deg, Illuminant = D65
+    X = var_R * 0.4124 + var_G * 0.3576 + var_B * 0.1805
+    Y = var_R * 0.2126 + var_G * 0.7152 + var_B * 0.0722
+    Z = var_R * 0.0193 + var_G * 0.1192 + var_B * 0.9505
+
+    # Convert XYZ to xy, see CIE 1931 color space on wikipedia
+    return X / (X + Y + Z), Y / (X + Y + Z)
